requirements.txt
Django==1.10.5
mysqlclient==1.3.9
django-debug-toolbar==1.6
Pillow==4.0.0
social-auth-app-django==1.1.0
python-social-auth==0.3.6
djangorestframework==3.5.4
django-oauth-toolkit==0.12.0
configparser==3.5.0
django-filter==1.0.1
